requirements.txt
pandas==2.2.3
pillow==11.1.0
torch==2.6.0
transformers==4.49.0
numpy==2.2.3
rembg==2.0.65
scikit-learn==1.6.1
pillow==11.1.0
matplotlib==3.10.1
onnxruntime==1.21.0
